+++ tests/xharness/Targets/TodayExtensionTarget.cs
@@ -68,7 +68,7 @@ namespace Xharness.Targets {
 				MonoNativeInfo.AddProjectDefines (csproj);
 				csproj.AddAdditionalDefines ("MONO_NATIVE_TODAY");
 			}
-			Harness.Save (csproj, TodayContainerProjectPath);
+			csproj.Save(TodayContainerProjectPath, Harness);
 
 			XmlDocument info_plist = new XmlDocument ();
 			var target_info_plist = Path.Combine ((IsGeneratedBclTest) ? GeneratedPath : TargetDirectory, $"Info{suffix}.plist");
@@ -76,7 +76,7 @@ namespace Xharness.Targets {
 			info_plist.SetCFBundleIdentifier (BundleIdentifier);
 			info_plist.SetCFBundleName (Name);
 			info_plist.SetMinimumOSVersion (GetMinimumOSVersion ("8.0"));
-			Harness.Save (info_plist, target_info_plist);
+			info_plist.Save (target_info_plist, Harness);
 		}
 
 		void CreateTodayExtensionProject ()
@@ -101,7 +101,7 @@ namespace Xharness.Targets {
 				csproj.AddAdditionalDefines ("MONO_NATIVE_TODAY");
 			}
 
-			Harness.Save (csproj, TodayExtensionProjectPath);
+			csproj.Save (TodayExtensionProjectPath, Harness);
 
 			TodayExtensionGuid = csproj.GetProjectGuid ();
 
@@ -122,7 +122,7 @@ namespace Xharness.Targets {
         <key>NSExtensionPointIdentifier</key>
         <string>com.apple.widget-extension</string>
     ");
-			Harness.Save (info_plist, target_info_plist);
+			info_plist.Save (target_info_plist, Harness);
 		}
 
 		protected override void ExecuteInternal ()
